LayoutTests/http/tests/mime/html-with-html-expected.txt

html-document.html has MIME type text/html
CONSOLE MESSAGE: Pass
